+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Example usage in Google Colab (requires `genbase>=0.2.17`)
+- `explabox.explore.instances` to summarize instances in a split
+- Selecting, sampling and filtering of `digestibles.Dataset`
 
 ### Fixed
 - Support for new pandas versions